Rename table to index
